feat(anchorperson): add related shows section
