+++ packages/mirror-tv-next/components/show/_slug/youtube-list-handler.tsx
@@ -2,7 +2,7 @@
 import useWindowDimensions from '~/hooks/use-window-dimensions'
 
 import YoutubeList from './youtube-list'
-import type { FormatPlayListItems } from './youtube-list'
+import type { FormatPlayListItems } from '~/types/api-data'
 import { useMemo, useState } from 'react'
 import styles from './_styles/youtube-list-handler.module.scss'
 import { fetchYoutubeList } from '../../../app/_actions/show-yt'
