+++ packages/mirror-tv-next/utils/youtube.ts
@@ -1,5 +1,5 @@
 import type { YoutubeItem, YoutubeResponse } from '~/types/youtube'
-import type { FormatPlayListItems } from '~/components/show/_slug/youtube-list'
+import type { FormatPlayListItems } from '~/types/api-data'
 import axios, { AxiosResponse } from 'axios'
 import { FetchError } from './index'
 import { YOUTUBE_API_URL } from '~/constants/environment-variables'
@@ -17,6 +17,7 @@ const formateYoutubeListRes = (
     }
   }
   return {
+    id: response?.id,
     items: filteredItems?.map((item) => formatPlayListItems(item)) ?? [],
     nextPageToken: response?.nextPageToken,
     totalItems: response.pageInfo.totalResults,
